requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
requests>=2.31.0
huggingface-hub>=0.16.0
kaggle>=1.5.16
